requirements.txt
python-docx
langchain
